+++ objects/src/transaction/mod.rs
@@ -20,7 +20,7 @@ pub use executed_tx::ExecutedTransaction;
 pub use inputs::{InputNote, InputNotes, ToNullifier, TransactionInputs};
 pub use outputs::{OutputNote, OutputNotes, TransactionOutputs};
 pub use prepared_tx::PreparedTransaction;
-pub use proven_tx::{AccountDetails, ProvenTransaction, ProvenTransactionBuilder};
+pub use proven_tx::{ProvenTransaction, ProvenTransactionBuilder};
 pub use transaction_id::TransactionId;
 pub use tx_args::{TransactionArgs, TransactionScript};
 pub use tx_witness::TransactionWitness;