+++ objects/src/transaction/proven_tx.rs
@@ -4,22 +4,17 @@ use miden_verifier::ExecutionProof;
 
 use super::{AccountId, Digest, InputNotes, Nullifier, OutputNote, OutputNotes, TransactionId};
 use crate::{
-    accounts::{Account, AccountDelta},
-    utils::serde::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
+    accounts::AccountDelta,
+    notes::{Note, NoteId},
+    utils::{
+        collections::*,
+        serde::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
+    },
     ProvenTransactionError,
 };
 
 // PROVEN TRANSACTION
 // ================================================================================================
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum AccountDetails {
-    /// The whole state is needed for new accounts
-    Full(Account),
-
-    /// For existing accounts, only the delta is needed.
-    Delta(AccountDelta),
-}
 
 /// Result of executing and proving a transaction. Contains all the data required to verify that a
 /// transaction was executed correctly.
@@ -41,7 +36,7 @@ pub struct ProvenTransaction {
 
     /// Optional account state changes used for on-chain accounts, This data is used to update an
     /// on-chain account's state after a local transaction execution.
-    account_details: Option<AccountDetails>,
+    account_delta: Option<AccountDelta>,
 
     /// A list of nullifiers for all notes consumed by the transaction.
     input_notes: InputNotes<Nullifier>,
@@ -83,9 +78,9 @@ impl ProvenTransaction {
         self.final_account_hash
     }
 
-    /// Returns the account details.
-    pub fn account_details(&self) -> Option<&AccountDetails> {
-        self.account_details.as_ref()
+    /// Returns the delta for on-chain accounts changes.
+    pub fn account_delta(&self) -> Option<&AccountDelta> {
+        self.account_delta.as_ref()
     }
 
     /// Returns a reference to the notes consumed by the transaction.
@@ -186,7 +181,7 @@ pub struct ProvenTransactionBuilder {
     final_account_hash: Digest,
 
     /// State changes to the account due to the transaction.
-    account_details: Option<AccountDetails>,
+    account_delta: Option<AccountDelta>,
 
     /// List of [Nullifier]s of all consumed notes by the transaction.
     input_notes: Vec<Nullifier>,
@@ -220,7 +215,7 @@ impl ProvenTransactionBuilder {
             account_id,
             initial_account_hash,
             final_account_hash,
-            account_details: None,
+            account_delta: None,
             input_notes: Vec::new(),
             output_notes: Vec::new(),
             tx_script_root: None,
@@ -232,9 +227,9 @@ impl ProvenTransactionBuilder {
     // PUBLIC ACCESSORS
     // --------------------------------------------------------------------------------------------
 
-    /// Sets the account's details.
-    pub fn account_details(mut self, account_details: AccountDetails) -> Self {
-        self.account_details = Some(account_details);
+    /// Sets the account's delta.
+    pub fn account_delta(mut self, account_delta: AccountDelta) -> Self {
+        self.account_delta = Some(account_delta);
         self
     }
 
@@ -266,15 +261,15 @@ impl ProvenTransactionBuilder {
     ///
     /// # Errors
     ///
-    /// An error will be returned if an on-chain account is used without provided on-chain detail.
-    /// Or if the account details, i.e. account id and final hash, don't match the transaction.
+    /// An error will be returned if an on-chain account is used without provided account delta.
     pub fn build(mut self) -> Result<ProvenTransaction, ProvenTransactionError> {
-        let account_details = self.account_details.take();
+        let account_delta = self.account_delta.take();
         let input_notes =
             InputNotes::new(self.input_notes).map_err(ProvenTransactionError::InputNotesError)?;
         let output_notes = OutputNotes::new(self.output_notes)
             .map_err(ProvenTransactionError::OutputNotesError)?;
         let tx_script_root = self.tx_script_root;
+
         let id = TransactionId::new(
             self.initial_account_hash,
             self.final_account_hash,
@@ -287,7 +282,7 @@ impl ProvenTransactionBuilder {
             account_id: self.account_id,
             initial_account_hash: self.initial_account_hash,
             final_account_hash: self.final_account_hash,
-            account_details,
+            account_delta,
             input_notes,
             output_notes,
             tx_script_root,
@@ -302,39 +297,12 @@ impl ProvenTransactionBuilder {
 // SERIALIZATION
 // ================================================================================================
 
-impl Serializable for AccountDetails {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        match self {
-            AccountDetails::Full(account) => {
-                0_u8.write_into(target);
-                account.write_into(target);
-            },
-            AccountDetails::Delta(delta) => {
-                1_u8.write_into(target);
-                delta.write_into(target);
-            },
-        }
-    }
-}
-
-impl Deserializable for AccountDetails {
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        match u8::read_from(source)? {
-            0_u8 => Ok(Self::Full(Account::read_from(source)?)),
-            1_u8 => Ok(Self::Delta(AccountDelta::read_from(source)?)),
-            v => Err(DeserializationError::InvalidValue(format!(
-                "Unknown variant {v} for AccountDetails"
-            ))),
-        }
-    }
-}
-
 impl Serializable for ProvenTransaction {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.account_id.write_into(target);
         self.initial_account_hash.write_into(target);
         self.final_account_hash.write_into(target);
-        self.account_details.write_into(target);
+        self.account_delta.write_into(target);
         self.input_notes.write_into(target);
         self.output_notes.write_into(target);
         self.tx_script_root.write_into(target);
@@ -348,7 +316,7 @@ impl Deserializable for ProvenTransaction {
         let account_id = AccountId::read_from(source)?;
         let initial_account_hash = Digest::read_from(source)?;
         let final_account_hash = Digest::read_from(source)?;
-        let account_details = <Option<AccountDetails>>::read_from(source)?;
+        let account_delta = <Option<AccountDelta>>::read_from(source)?;
 
         let input_notes = InputNotes::<Nullifier>::read_from(source)?;
         let output_notes = OutputNotes::read_from(source)?;
@@ -370,7 +338,7 @@ impl Deserializable for ProvenTransaction {
             account_id,
             initial_account_hash,
             final_account_hash,
-            account_details,
+            account_delta,
             input_notes,
             output_notes,
             tx_script_root,